feat: Exchange Request/Response API - Remove receiver and add Transport to function args
Currently, the SendExchangeRequest and SendExchangeResponse are part of DIDComm receiver/member function. The functions use the receiver variable Transport, which is injected to the DIDComm receiver. Instead, transport can be passed as an argument to these functions. This way lib consumer don't have to maintain/use DIDComm instance to invoke these functions(plus will result in better readability).

Closes #20

Signed-off-by: Rolson Quadras <rolson.quadras@securekey.com>