refactor: Update did-comm-go references to aries-framework-go
Aries spec implementation was divided into did-comm-go and agent-framework-go repos in https://github.com/trustbloc. Going forward these will be part of one repo and as such, the did-comm-go repo was renamed into aries-framework-go. The codebase references need to be updated to be in sync with the repo name change.

Closes #27

Signed-off-by: Rolson Quadras <rolson.quadras@securekey.com>